♻️ [refact] Quiz 생성 로직 리팩터링 - 실제 퀴즈 생성 로직 이동
